fix: translate by moving pages into client folder so there is only one path to translate.js so it gets uploaded on the browser only once
